backend/requirements.txt
flask
redis
requests
flask_caching
python-dotenv